merge: 민지 영훈 member domain 합치기
