+++ src/main/java/com/fourback/runus/domains/member/controller/SuccessController.java
@@ -1,4 +1,4 @@
-package com.fourback.runus.domains.members.controller;
+package com.fourback.runus.domains.member.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
