+++ src/main/java/com/fourback/runus/domains/member/service/JpaUserDetailsService.java
@@ -1,13 +1,13 @@
-package com.fourback.runus.domains.members.service;
+package com.fourback.runus.domains.member.service;
 
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
-import com.fourback.runus.domains.members.domain.Member;
-import com.fourback.runus.domains.members.dto.CustomUserDetails;
-import com.fourback.runus.domains.members.repository.MemberRepository;
+import com.fourback.runus.domains.member.domain.Member;
+import com.fourback.runus.domains.member.dto.CustomUserDetails;
+import com.fourback.runus.domains.member.repository.MemberRepository;
 
 import lombok.RequiredArgsConstructor;
