--- src/main/java/com/fourback/runus/domains/members/domain/Member.java
@@ -1,58 +1,0 @@
-package com.fourback.runus.domains.members.domain;
-
-import java.math.BigDecimal;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "members")
-@Data
-@AllArgsConstructor @NoArgsConstructor @Builder
-public class Member {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long userId;
-
-    @Column(nullable = false, unique = true)
-    private String email;
-
-    @Column(nullable = false, unique = true)
-    private String nickName;
-
-    @Column(nullable = false)
-    private String password;
-
-    @Column(nullable = false)
-    private LocalDate birth;
-
-    private String profileUrl;
-
-    @Column(nullable = false)
-    private BigDecimal height;
-
-    @Column(nullable = false)
-    private BigDecimal weight;
-
-    @Column(nullable = false)
-    private String role = "USER";
-
-    @Column(nullable = false, updatable = false)
-    private LocalDateTime createdAt = LocalDateTime.now();
-
-    @Column(nullable = false)
-    private LocalDateTime updatedAt;
-
-    private LocalDateTime deletedAt;
-}
