+++ src/main/java/com/fourback/runus/global/security/provider/JwtProvider.java
@@ -12,7 +12,7 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Component;
 
-import com.fourback.runus.domains.members.service.JpaUserDetailsService;
+import com.fourback.runus.domains.member.service.JpaUserDetailsService;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
